Placed Upload Image function in separate file
