+++ Controllers/EventControllers/EventCreation.js
@@ -1,13 +1,8 @@
-const cloudinary = require('cloudinary').v2;
+// const cloudinary = require('cloudinary').v2;
 const { createEvent } = require('../../SQL/EventQueries/CreateEvent');
 const { findByAttribute } = require('../../SQL/AuthQueries/FindExistingEntity');
 const { addEventCategories } = require('../../SQL/EventQueries/AddEventCategories');
-
-cloudinary.config({
-    cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
-    api_key: process.env.CLOUDINARY_API_KEY,
-    api_secret: process.env.CLOUDINARY_API_SECRET
-});
+const uploadImageToCloudinary = require('../../Util/UploadImage');
 
 module.exports.EventCreation = async (req, res) => {
     try {
@@ -27,30 +22,6 @@ module.exports.EventCreation = async (req, res) => {
             res.status(400).send({ message: 'Image file is required' });
             return;
         }
-
-        const uploadImageToCloudinary = (image) => {
-            return new Promise((resolve, reject) => {
-                const { originalname, mimetype, buffer } = image;
-
-                cloudinary.uploader.upload_stream({ resource_type: 'image' }, (error, result) => {
-                    if (error) return reject(error);
-
-                    const { public_id, secure_url } = result;
-                    const url = cloudinary.url(public_id, {
-                        width: 650,
-                        height: 450,
-                        crop: "fill"
-                    });
-
-                    resolve({
-                        name: originalname,
-                        type: mimetype,
-                        url: secure_url,
-                        public_id: public_id
-                    });
-                }).end(buffer);
-            });
-        };
 
         const uploadedImageData = await uploadImageToCloudinary(image);
